Fix missing dependency error
Like travis has here https://travis-ci.org/tngan/samlify/jobs/401602020
I have encountered an error arround the xml-encryption library. For some
reason it seems to have been missing. Adding the prefix `@passify/'
seems to fix it.

This is not related to the current PR and you might want extract this
commit if the rest of the PR fails for some reason.