Expose InResponseTo in parseLoginResponse extract (#190)
* Fix missing dependency error

Like travis has here https://travis-ci.org/tngan/samlify/jobs/401602020
I have encountered an error arround the xml-encryption library. For some
reason it seems to have been missing. Adding the prefix `@passify/'
seems to fix it.

This is not related to the current PR and you might want extract this
commit if the rest of the PR fails for some reason.

* Return InResponseTo field from parseLoginResponse

This seems to be returning the right field and should allow the users to
check that the id matches the nonce that was send during sp initiated
logins.

I've updated the exsiting tests rather than creating a new one because
this is not really a new use case.