fix test: remove sessionIndex from failing logout request tests
Seems to have been accidentally added in https://github.com/tngan/samlify/pull/501,
these never worked because `sessionIndex` is not part of the default logout request template,
so whether the extractor works or not, the creation of the request will not include the value.

And with this, the test suite passes again!